Changing logging prints to use long long ints for time (like they should) instead of long ints.

+++ src/kernel/Logger.c
@@ -89,7 +89,7 @@ const char _logLevelNames[NUM_LOG_LEVELS][9] KEEP_IN_FLASH = {
 /// @note KEEP_IN_FLASH is required here because .rodata is removed from the
 /// final binary on some targets.
 static const char _logHeaderFormat[] KEEP_IN_FLASH
-  = "[%ld.%09ld %s:%u %s:%u %s] ";
+  = "[%lld.%09lld %s:%u %s:%u %s] ";
 
 /// @var _localhost
 ///
@@ -234,8 +234,8 @@ writeImmediate:
   
   snprintf(HAL->memory.logBuffer, HAL->memory.logBufferSize,
     _logHeaderFormat,
-    (long int) (commandArgs.logEntry.timeStamp / ((int64_t) 1000000000)),
-    (long int) (commandArgs.logEntry.timeStamp % ((int64_t) 1000000000)),
+    (long long int) (commandArgs.logEntry.timeStamp / ((int64_t) 1000000000)),
+    (long long int) (commandArgs.logEntry.timeStamp % ((int64_t) 1000000000)),
     ((SCHEDULER_STATE != NULL) && (SCHEDULER_STATE->hostname != NULL))
       ? SCHEDULER_STATE->hostname
       : _localhost,
